Re-factor schema class and configuration.
Preparing for one more backend type and want to make things reusable
between backends. Factored out piece of schema class that does not
depend on sqlalchemy and moved it into a new class PpdbBaseSchema.
PpdbSchema renamed to PpdbSqlSchema and inherits from PpdbBaseSchema.
PpdbConfig was also split into few classes corresponding to schema
classes.
